layers: Move Device CoreCheck into own file
